// modules/transfer/modules/transfer-third/enums/third-transfer-control-name.enum.ts
export enum AttributeFormCrud {
  TYPE_ACCOUNT = 'typeAccount',
  NUMBER_ACCOUNT = 'numberAccount',
  NAME_ACCOUNT = 'nameAccount',
  ACCOUNT_STATE = 'accountState',
  AMOUNT = 'amount',
  COMMENT = 'comment',
  CURRENCY = 'CURRENCY',
  ALIAS = 'alias',
  EMAIL = 'email',
}

export enum AttributeThirdFormTransfer {
  TYPE_ACCOUNT = 'typeAccount',
  NUMBER_ACCOUNT = 'numberAccount',
  NAME_ACCOUNT = 'nameAccount',
  ACCOUNT_STATE = 'accountState',
  AMOUNT = 'amount',
  COMMENT = 'comment',
  CURRENCY = 'currency',
  ACCOUNT_ACCREDIT_CURRENCY = 'accountAccreditCurrency',
  ACCOUNT_ACCREDIT_ALIAS = 'accountAccreditAlias',
  ACCOUNT_ACCREDIT_TYPE = 'accountAccreditType',
  DESTINATION_CURRENCY = 'destinationCurrency',
  DESCRIPTION_CURRENCY = 'descriptionCurrency',
  ACCOUNT_DEBITED = 'accountDebited',
  ACCOUNT_ACCREDIT = 'accountAccredit',
  EMAIL = 'email',
  ALIAS = 'alias',
}

export enum AttributeThirdTransferTable {
  FILTER = 'filter',
  SELECT_TYPE_FILTER = 'typeFilter',
}